wifi: Update SimpleWiFi.py to support page view for AP list
This patch enables page view function for AP list. When it scans
more number of access points than vertical size of UI, it can
show all the access points in page view way. For instance, if
the UI allows 15 qty in a page and 33 APs found, this provides
3 pages of access point list so that users can navigate the list
with 'n' or 'p' for next or previous, respectively.

Change-Id: I76031be320cb8034ca39fd21fb5c43eea529657b
Signed-off-by: Siwon Kang <sw0809.kang@samsung.com>